Improve the container registry image tags list
This uses the new image-tags endpoint in platform-api, to get a list of
tags (sorted by creation date), with additional meta data.
It also adds linking between the tags list and the create microservice
screen, where it prefils the head image field with the value found in
the fragment identifier.