Remove manifest.json as we don't need it
It is used to build PWAs - and I don't think we need it.
The way it was setup now it slowed down the initial page load because it
blocked the dom content loading while it was fetching the file. Chrome
also seems to load this file without cookies - which caused it to be
redirected to login.

If we want to reintroduce this file - we need to coordinate some changes
in the platform ingress setup.